add receipt list view by user role and implement method getAllUserReceiptsByUserId

+++ src/main/java/db/dao/interfaces/ReceiptDao.java
@@ -3,14 +3,13 @@ package db.dao.interfaces;
 import entities.Receipt;
 import exceptions.DatabaseException;
 
-import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.List;
 
 public interface ReceiptDao {
     boolean addReceipt(Receipt receipt) throws SQLException;
 
-    List<Receipt> getAllUserReceiptsByUserId(Long id) throws DatabaseException;
+    List<Receipt> getAllUserReceiptsByUserId(int id) throws DatabaseException;
 
     List<Receipt> getAllReceipts() throws DatabaseException;
